Develop

+++ filewatcher/server_class.py
@@ -15,6 +15,7 @@ from filewatcher.utils import (
     download_file,
     download_folder,
     get_files,
+    get_folders,
     read_data,
 )
 
@@ -65,6 +66,7 @@ class ServerFwr:
                 'err': 'Invalid password'
             }).encode('utf-8')
         log.debug("%s %s", command, args)
+
         res = None
         error = None
         if command == Commands.SHOW_FOLDER.name:
@@ -79,6 +81,13 @@ class ServerFwr:
             res, error = self.delete(args)
         elif command == Commands.CHECK_THREE.name:
             res, error = self.check_three(args)
+        elif command == Commands.RENAME.name:
+            res, error = self.rename(args)
+        elif command == Commands.MOVE.name:
+            res, error = self.move(args)
+        elif command == Commands.COPY.name:
+            res, error = self.copy(args)
+
         if res is None and error is None:
             return
         return dumps({
@@ -116,6 +125,8 @@ class ServerFwr:
         self.socket.close()
 
     def download(self, folder: str):
+        if '..' in folder:
+            return 0, "Invalid path"
         error = None
         download_path = os.path.join(self.directory, folder)
         path, name = os.path.split(folder)
@@ -135,12 +146,15 @@ class ServerFwr:
     def upload(self, path_info: dict):
         if path_info.get('isfile'):
             size, path, filename = path_info.get('size'), path_info.get('path'), path_info.get('filename')
-            if not os.path.isdir(os.path.join(self.directory, path)):
+            if not os.path.isdir(os.path.join(self.directory, path)) or '..' in path:
                 return None, "Invalid path {}".format(path)
             download_file(self.connection, size, os.path.join(self.directory, path, filename))
             return 1, None
         elif path_info.get('isfolder'):
-            foldername, path, count_files = path_info.get('foldername'), path_info.get('path'), path_info.get('count_files')
+            foldername, path, count_files = path_info.get('foldername'), path_info.get('path'), path_info.get(
+                'count_files')
+            if '..' in path:
+                return 0, "Invalid path {}".format(path)
             if path == '.':
                 path = ''
 
@@ -149,9 +163,14 @@ class ServerFwr:
         return None, "Invalid path info {}".format(path_info)
 
     def delete(self, delete_dir: str):
+        if '..' in delete_dir:
+            return 0, "Invalid path"
         if not delete_dir or delete_dir.startswith('/'):
             return 0, "Invalid path"
-
+        if delete_dir == '.':
+            shutil.rmtree(self.directory)
+            os.makedirs(self.directory)
+            return 1, None
         delete_dir = os.path.join(self.directory, delete_dir)
         if os.path.isfile(delete_dir):
             os.remove(delete_dir)
@@ -161,14 +180,94 @@ class ServerFwr:
             return 0, "Invalid path"
         return 1, None
 
-    def check_three(self, three: list):
-        this_three = list(get_files(self.directory, True))
+    def check_three(self, three: dict):
+        files_three = three.get('files')
+        folders_three = three.get('folders')
 
-        delete_dirs = [d for d in this_three if d not in three]
-        need_dirs = [d for d in three if d not in this_three]
+        this_folders_three = list(get_folders(self.directory, is_root=True))
+        for folder in this_folders_three:
+            if folder not in folders_three:
+                shutil.rmtree(os.path.join(self.directory, folder))
 
-        for directory in delete_dirs:
+        for folder in folders_three:
+            if folder not in this_folders_three:
+                os.makedirs(os.path.join(self.directory, folder))
+
+        this_files_three = list(get_files(self.directory, is_root=True, get_size=True))
+        for directory in [d[0] for d in this_files_three if d not in files_three]:
             directory = os.path.join(self.directory, directory)
             if os.path.isfile(directory):
                 os.remove(directory)
-        return need_dirs, None
+
+        return [d[0] for d in files_three if d not in this_files_three], None
+
+    def rename(self, rename_d: list):
+        old_name, new_name = rename_d
+        if ".." in old_name:
+            return 0, "Invalid path"
+
+        if not os.path.exists(os.path.join(self.directory, old_name)):
+            return 0, "{} not exist".format(old_name)
+        if os.path.exists(os.path.join(self.directory, new_name)):
+            return 0, "{} name already exist".format(new_name)
+
+        os.renames(os.path.join(self.directory, old_name), os.path.join(self.directory, new_name))
+        return 1, None
+
+    def move(self, move_d: list):
+        object_move, move_destination = move_d
+        if object_move == '.' or '..' in object_move:
+            return 0, "Invalid path"
+
+        obj_move = os.path.join(self.directory, object_move)
+        dst_move = os.path.join(self.directory, move_destination)
+
+        if not os.path.exists(obj_move):
+            return 0, "Invalid source path"
+
+        if dst_move.endswith('/'):
+            if not os.path.exists(dst_move):
+                os.makedirs(dst_move)
+            elif os.path.split(object_move)[1] in os.listdir(dst_move):
+                return 0, "{} already exist".format(os.path.join(move_destination, os.path.split(object_move)[1]))
+        else:
+            path, name = os.path.split(dst_move)
+            if name == '.':
+                name = os.path.split(object_move)[1]
+
+            if not os.path.exists(path):
+                os.makedirs(path)
+            elif name in os.listdir(path):
+                return 0, "{} already exist".format(os.path.join(os.path.split(move_destination)[0], name))
+
+        shutil.move(obj_move, dst_move)
+        return 1, None
+
+    def copy(self, copy_d: list):
+        copy_dir, copy_dst = copy_d
+        if copy_dir == '.' or '..' in copy_dir:
+            return 0, "Invalid path"
+
+        copy_directory = os.path.join(self.directory, copy_dir)
+        copy_destination = os.path.join(self.directory, copy_dst)
+
+        if not os.path.exists(copy_directory):
+            return 0, "{} doesn't exist".format(copy_dir)
+
+        if copy_destination.endswith('/'):
+            if not os.path.exists(copy_destination):
+                os.makedirs(copy_destination)
+            elif os.path.split(copy_dir)[1] in os.listdir(copy_destination):
+                return 0, "{} already exist".format(os.path.join(copy_dst, copy_dir))
+        else:
+            path, name = os.path.split(copy_destination)
+            if name == '.':
+                name = os.path.split(copy_dir)[1]
+
+            if not os.path.exists(path):
+                os.makedirs(path)
+            elif name in os.listdir(path):
+                return 0, "{} already exist".format(os.path.join(os.path.split(copy_dst)[0], name))
+
+        shutil.copy(copy_directory, copy_destination)
+        return 1, None
